check-engine.js -> check.js and moved to root
